utils/proto: consolidate proto helpers
Consolidates helpers into a separate file in the
utils pkg for reusability. The proto marshalling
to YAML will be required in a subsequent change
that builds the envoy bootstrap config in a generic
way for code reuse between the injector and OSM
gateway.
